# twitter-swagger.yaml
openapi: 3.0.4
info:
  title: Twitter clone TypeScript - 2025
  description: |-
    This is a sample Pet Store Server based on the OpenAPI 3.0 specification.  You can find out more about
    Swagger at [https://swagger.io](https://swagger.io). In the third iteration of the pet store, we've switched to the design first approach!
    You can now help us improve the API whether it's by making changes to the definition itself or to the code.
    That way, with time, we can improve the API in general, and expose some of the new features in OAS3.

    Some useful links:
    - [The Pet Store repository](https://github.com/swagger-api/swagger-petstore)
    - [The source API definition for the Pet Store](https://github.com/swagger-api/swagger-petstore/blob/master/src/main/resources/openapi.yaml)
  termsOfService: https://swagger.io/terms/
  contact:
    email: ngpham.2807@gmail.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.12
...
servers:
  - url: http://localhost:4000
    description: Local server
...
paths:  
...
  /users/refresh-token:
    post:
      tags:
        - users
      summary: Refreshes the access token
      operationId: refreshToken
      requestBody:
        description: The refresh token to be used for generating a new access token
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refresh_token:
                  type: string
                  example: 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjoiNjg0NmE5MjJjOGQ2YzYwYzI5ZWM5MmQzIiwidG9rZW5fdHlwZSI6MSwidmVyaWZ5IjoxLCJpYXQiOjE3NTI1ODk5MzYsImV4cCI6MTc2MTIyOTkzNn0.C4W_Me1Di-owhrpj1VJ6j-ImoPtWMy6QMPT6R-uJ4a8'
      responses:
        '200':
          description: Get new access token
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: New tokens returned successfully
                  result:
                    $ref: '#/components/schemas/AuthenticateSuccessfully'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Used refresh token or does not exist'
...
  /users/verify-email:
    post:
      tags:
        - users
      summary: Verify user email through SES
      operationId: verifyEmail
      requestBody:
        description: The email verification token
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email_verify_token:
                  type: string
                  example: 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjoiNjg4MTAyYjcwN2RlYjI1MWE4MzAxYjY0IiwidG9rZW5fdHlwZSI6MywidmVyaWZ5IjowLCJpYXQiOjE3NTMyODUzMDMsImV4cCI6MTc1Mzg5MDEwM30.-LafPF4V9FDHU6Ondluf5C8Gc3yh8O9pQTEn9HXanmE'
      responses:
        '200':
          description: Email verified successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Email verified successfully'
                  result:
                    $ref: '#/components/schemas/AuthenticateSuccessfully'

        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Email verify token is required'
...
  /users/forgot-password:
    post:
      tags:
        - users
      summary: Request password reset
      operationId: forgotPassword
      requestBody:
        description: The email to send the password reset token
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: 'ngpham81@gmail.com'
      responses:
        '200':
          description: Password reset token sent successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  result:
                    type: object
                    properties:
                      message:
                        type: string
                        example: 'Check your email for the password reset token'
                      forgot_password_token:
                        type: string
                        example: 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjoiNjg0NmE5MjJjOGQ2YzYwYzI5ZWM5MmQzIiwidG9rZW5fdHlwZSI6MiwidmVyaWZ5IjoxLCJpYXQiOjE3NTMzMzA3MDgsImV4cCI6MTc1MzQxNzEwOH0.HV6PAzBoEM6vmQ3o4qXSsC_BBJVsO6ws4SUa7CvobhM'
        '422':
          description: Unprocessable Entity
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Validation error'
                  errors:
                    type: object
                    properties:
                      email:
                        type: object
                        properties:
                          type:
                            type: string
                            example: 'field'
                          value:
                            type: string
                            example: 'Non-existent-Email@gmail.com'
                          msg:
                            type: string
                            example: 'User not found'
                          path:
                            type: string
                            example: 'email'
                          location:
                            type: string
                            example: 'body'
...
  /users/reset-password:
    post:
      tags:
        - users
      summary: Send forgot password token through SES
      operationId: resetPassword
      requestBody:
        description: The new password and confirmation
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                forgot_password_token:
                  type: string
                  example: 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjoiNjg0NmE5MjJjOGQ2YzYwYzI5ZWM5MmQzIiwidG9rZW5fdHlwZSI6MiwidmVyaWZ5IjoxLCJpYXQiOjE3NTMzMzIzMjEsImV4cCI6MTc1MzQxODcyMX0.C_nF9pAdRvQzdQvDn2k_hxGxuElq_qohwgVVOtQmhw4'
                password:
                  type: string
                  example: 'NewPassword123!'
                confirm_password:
                  type: string
                  example: 'NewPassword123!'
      responses:
        '200':
          description: Password reset successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  result:
                    type: object
                    properties:
                      message:
                        type: string
                        example: 'Reset password successfully'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Invalid forgot password token'
        '422':
          description: Unprocessable Entity
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Validation error'
                  errors:
                    type: object
                    properties:
                      confirm_password:
                        type: object
                        properties:
                          type:
                            type: string
                            example: 'field'
                          value:
                            type: string
                            example: 'NewPassword123!'
                          msg:
                            type: string
                            example: 'Password confirmation does not match password'
                          path:
                            type: string
                            example: 'confirm_password'
                          location:
                            type: string
                            example: 'body'
...
  /users/change-password:
    put:
      tags:
        - users
      summary: Change user password
      operationId: changePassword
      security:
        - BearerAuth: []
      requestBody:
        description: The current password and new password
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                old_password:
                  type: string
                  example: 'CurrentPassword123!'
                new_password:
                  type: string
                  example: 'NewPassword123!'
                confirm_new_password:
                  type: string
                  example: 'NewPassword123!'
      responses:
        '200':
          description: Password changed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  result:
                    type: object
                    properties:
                      message:
                        type: string
                        example: 'Change password successfully'
        '422':
          description: Unprocessable Entity
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Validation error
                  errors:
                    type: object
                    properties:
                      confirm_new_password:
                        type: object
                        properties:
                          type:
                            type: string
                            example: field
                          value:
                            type: string
                            example: Phuong2872@@!
                          msg:
                            type: string
                            example: Password confirmation does not match password
                          path:
                            type: string
                            example: confirm_new_password
                          location:
                            type: string
                            example: body
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Old password does not match'
              examples:
                oldPasswordMismatch:
                  summary: Old password mismatch
                  value:
                    message: 'Old password does not match'
                accessTokenRequired:
                  summary: Access token required
                  value:
                    message: 'Access token is required'
...
components:
  schemas:
...
    AuthenticateSuccessfully:
      type: object
      properties:
        result:
          type: object
          properties:
            access_token:
              type: string
              example: 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjoiNjg0NmE5MjJjOGQ2YzYwYzI5ZWM5MmQzIiwidG9rZW5fdHlwZSI6MCwidmVyaWZ5IjoxLCJpYXQiOjE3NTI1ODk5MzYsImV4cCI6MTc1MjU4OTk2Nn0.N0NpQZDcu-gMpXLApVimd0etDPJoz7cp6ksGsdbP1WA'
            refresh_token:
              type: string
              example: 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjoiNjg0NmE5MjJjOGQ2YzYwYzI5ZWM5MmQzIiwidG9rZW5fdHlwZSI6MSwidmVyaWZ5IjoxLCJpYXQiOjE3NTI1ODk5MzYsImV4cCI6MTc2MTIyOTkzNn0.C4W_Me1Di-owhrpj1VJ6j-ImoPtWMy6QMPT6R-uJ4a8'
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Use the `Bearer` scheme to authenticate requests. Include the token in the `Authorization` header as follows:
        ```
        Authorization: Bearer <your_token>
        ```